requirements.txt
django>=4.2.2
gunicorn>=21.2.0
django-debug-toolbar>=4.2.0
django-environ

# Testing tools
pytest-django>=4.7.0
pytest-dotenv>=0.5.2
pytest-factoryboy>=2.6.0
faker>=21.0.0

whitenoise
django-theme-soft-design==1.0.10

python-decouple==3.6
psycopg2-binary>=2.9.6
redis>=4.1
hiredis>=2.2.3
celery>=5.2
flower>=1.2.0
django-bootstrap-v5>=1.0.11
# misaka>=2.1.1
django-braces>=1.14.0
Pillow>=10.1.0
django-celery-results>=2.5.1
jsonschema>=4.20.0

# Email Backend
django-ses>=3.5.0
django-ses[events]

# psycopg2-binary
# mysqlclient

# Better json widget
django-json-widget>= 1.1.1
